appraisal, don't validate points for self-appraisal

+++ patches/january_2013/change_patch_structure.py
@@ -2,11 +2,12 @@ import webnotes
 
 def execute():
 	webnotes.reload_doc("core", "doctype", "patch_log")
-	for d in webnotes.conn.sql("""select patch from __PatchLog"""):
-		webnotes.doc({
-			"doctype": "Patch Log",
-			"patch": d[0]
-		}).insert()
+	if webnotes.conn.table_exists("__PatchLog"):
+		for d in webnotes.conn.sql("""select patch from __PatchLog"""):
+			webnotes.doc({
+				"doctype": "Patch Log",
+				"patch": d[0]
+			}).insert()
 	
-	webnotes.conn.commit()
-	webnotes.conn.sql("drop table __PatchLog")
+		webnotes.conn.commit()
+		webnotes.conn.sql("drop table __PatchLog")
